+++ examples/components/dropdown-menu-view/dropdown-menu-view.component.jsx
@@ -1,7 +1,9 @@
 /* eslint-disable no-console */
 
 import React from 'react';
+
 import { DropdownMenu } from '../../../src/index';
+import { Icon } from '../../../src/icons';
 
 function DropdownMenuView() {
   const containerStyle = {
@@ -19,61 +21,71 @@ function DropdownMenuView() {
     background: 'white',
   };
 
+  const menuItems1 = [
+    {
+      id: 'item_id_11',
+      title: 'Item 1, dont\'t close',
+      onClick: () => console.log('Item 1 clicked'),
+      disableClosing: true,
+    },
+    {
+      id: 'item_id_12',
+      title: 'Item 2, with the icon',
+      onClick: () => console.log('Item 2 clicked'),
+      icon: <Icon type="indicator" name="ok" width={25} height={25} />,
+    },
+    {
+      id: 'item_id_d1',
+      type: 'divider',
+    },
+    {
+      id: 'item_id_13',
+      title: 'Item 3',
+      onClick: () => console.log('Item 3 clicked'),
+      disabled: true,
+    },
+  ];
+
+  const menuItems2 = [
+    {
+      id: 'item_id_21',
+      title: 'Item 1',
+      onClick: () => console.log('Item 1 clicked'),
+    },
+    {
+      id: 'item_id_22',
+      title: 'Item 2',
+      onClick: () => console.log('Item 2 clicked'),
+    },
+    {
+      id: 'item_id_d1',
+      type: 'divider',
+    },
+    {
+      id: 'item_id_23',
+      title: 'Item 3',
+      onClick: () => console.log('Item 3 clicked'),
+      disableClosing: true,
+    },
+  ];
+
   return (
     <div>
       <div style={containerStyle}>
         {/* Default dropdown menu with three dots */}
         <DropdownMenu
           id="example"
-          menuItems={[
-            {
-              id: 'item_id_1',
-              title: 'Item 1, dont\'t close',
-              disableClosing: true,
-              onClick: () => console.log('Item 1 clicked'),
-            },
-            {
-              title: 'Item 2',
-              icon: <i className="fa fa-info" />,
-              onClick: () => console.log('Item 2 clicked'),
-            },
-            {
-              type: 'divider',
-            },
-            {
-              title: 'Item 3',
-              disabled: true,
-              onClick: () => console.log('Item 3 clicked'),
-            },
-          ]}
+          menuItems={menuItems1}
         />
       </div>
       <div style={containerStyle2}>
         {/* More customized dropdown menu */}
         <DropdownMenu
           id="example2"
+          menuItems={menuItems2}
           title="Dropdown"
-          pullRight={false}
           caret
-          menuItems={[
-            {
-              title: 'Item 1',
-              onClick: () => console.log('Item 1 clicked'),
-            },
-            {
-              title: 'Item 2',
-              icon: <i className="fa fa-info" />,
-              onClick: () => console.log('Item 2 clicked'),
-            },
-            {
-              type: 'divider',
-            },
-            {
-              title: 'Item 3',
-              disabled: true,
-              onClick: () => console.log('Item 3 clicked'),
-            },
-          ]}
+          pullRight={false}
         />
       </div>
     </div>