+++ CHANGELOG.md
@@ -2,19 +2,23 @@
 
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
-## <next>
+## 3.0.0
 
+* [npm scope](https://docs.npmjs.com/misc/scope) `@opuscapita` added to the package name
+* Package set to public
 * `Cards` fix collapse/expand icon container placement
 * `Menu` component added
 * `KEY_CODES` constant added
 * `PerfProfiler` utility component added
+* `DropdownMenu`, `MultiSelect`, and `MultiSelectItem` components refactored
+* `DropdownContainer`, `DropdownMultiSelect` components added
 
 ## 2.0.1
 
 * `ResponsiveNavbar` add timeout to componentDidMount to wait for real rendering
 * `ApplicationLayout` disabled touch for sidebar: drag area on left is reserved by browser e.g. in iPad, sidebar burger icon fill is white.
 * npm script `examples` is now `docs` and related files are renamed accordingly
-* `MultiSelect` and `MultiSelectItem` components
+* `MultiSelect` and `MultiSelectItem` components added
 
 ## 2.0.0